Fixes: ST: allow the same vmi to be referenced in multiple service-instances
The VirtualMachineInterfaceST class has a singular variable to store port_tuple.
Making it an array to keep updating if there are multiple port tuples.
Also, having multiple port tuples will work only in the case of in-network
and in-network-net service mode. Also, adding a UT to check for port tuples.

Change-Id: Ie57796d1b4edf362072c4c5e5b043754f4e374fe
Closes-Bug: 1661134